fix: the check for ``php` is not recognized` error being too specific as flagged by copilot.
The original solution only checked for the literal string ``php` is not recognized` but that error message is only on Windows CMD and Powershell.

To fix this and make it OS and shell agnostic, we need to construct a regex and test that instead.

The regex tests for the 5 common formats for the "command not found" error. It also works for with and without single or double quotes around `php`. Also with or without a colon (`:`).

The 5 common formats (without quotes) are:
1. php is not recognized
2. php: command not found
3. php cannot be found
4. command not found: php
5. Unknown command php

While this works for these 5, they're may be some edge cases where certain OS or shells format the error differently and therefore the regex fails. Add those edge cases to the regex as and when they occur.