fix: on-hover intellisense diagnostic source on same line as icon.
In the editor, the on-hover intellisense tooltip would have render the diagnostic source (`phpcs`) on the same line as the icon. This made it look weird.

Fixed by added a new line before the diagnostic source. This new line only renders in the intellisense tool tip and not in the Problems panel.