Feat: Display icons on problems to determine their auto fixability (#165)
* feat: display icons on errors in the problems tab

- Allow the `Diagnostic` constructor to show whether each error is auto fixable or not by appending the new message to the output. This will make the errors more readable in the Problems tab.

We're already setting the `fixable` as `boolean` per error in the `PHPCSMessage` interface, so we can just use this to determine the icon. We have to use copy/paste-able character icons (✔️❌)  since the API doesn't allow codicons.

* fix: on-hover intellisense diagnostic source on same line as icon.

In the editor, the on-hover intellisense tooltip would have render the diagnostic source (`phpcs`) on the same line as the icon. This made it look weird.

Fixed by added a new line before the diagnostic source. This new line only renders in the intellisense tool tip and not in the Problems panel.