Improve Error Reporting (#171)
* feat: log the exit code, stdout, stderr, and node error to Output.

For better debugging and error reporting, we now log all output streams to the Output channel when they are not empty.

* fix: ts errors

ts errors via github actions: https://github.com/valeryan/vscode-phpsab/actions/runs/18027596186/job/51297731350?pr=171

* fix: fixer logging of `STDOUT` only if it contains the string `ERROR`.

- Added a check to see if `STDOUT` starts with the string `ERROR`. We only log it if it does. Also added a comment to document the reason behind it.

* refactor: remove duplicate debug message logging in fixer `format`

- Removed duplicate `showInformationMessage` calls in the fixer's format switch-case, and added 1 singular call at the bottom of the function.

- Removed the conditional debugging for showing the information messages to the user since info like "fixed the errors" etc. should always be shown to the users for better UX.

- Since case 0 is only logging the message that is already set before the switch-case, then we can now remove this case, as the message will be logged at the bottom of the function.

* fix: fixer general error reporting

- Change the multiple lone if statements to if...else if...else for improved readability since errors can't have multiple error codes, as it's one or the other. Also added a catch all else statement to log all other errors that the if statements didn't catch.

* refactor: fixer error logging and add message logging.

- Removed the logging of the `fixed` variable.

- Added logging of the errors at the bottom of the `format` function.

- Added logging of the fixer message if there are no errors. This logs the messages such as "fixed all errors", to the Output channel for better debugging.

* revert: PR #58 which introduced error thrown when phpcbf returns empty.

While the fix was to throw an error when phpcbf returned an empty document (text variable is an empty string), which prevented the file from being emptied. But this thrown error "PHPCBF returned an empty document" isn't necessarily descriptive enough for debugging. It relates to the file not needing formatting changes, which is already a message in the form of "No fixable errors were found".

Because PR #160 fixes multiple issues with the promise logic, Copilot recommended "just resolving with no edits rather than throwing an error." This would also promote a better UX.

- Reverted back to the just resolving with no edits instead of throwing an error.

Note: This doesn't cause the original bug to come back and remove a file's content.

* remove: old redundant vscode `editor.formatOnSaveTimeout` setting.

The `editor.formatOnSaveTimeout` setting was removed from vscode v1.42 in 2020. PR #15 removed the timeout dependency from the extension. But this one seemed to have been missed. Removing since it doesn't do anything now.

* refactor: improve node errors in fixer.

- Refactored the `null` switch-case, (which deals with specific Node errors, eg. `ENOENT`) into a new function: `determineNodeError`.

This new function determine's the Node error from it's exit code. It uses Node's `getSystemErrorMap` util function to get a Map of all system error codes  and their description's (like `ENOENT`), and we can more precisely generate a better error message. This util function only maps the system errors though, and not the internal Node errors that start with `ERR_`. So we still need to map the `ERR_OPERATION_FAILED` code to it's message.

- Added new `cause` property to the `ConsoleError` interface because if an error is caught and then a new error is thrown with a different message/code, we still need to be able to see the originating error. So `cause` will be the originating error. It's probably very unlikely this will happen, but on the off chance it does, then debugging would be easier having access to the original error.

- Added new `path` property to the `ConsoleError` interface since Node catches system errors in it's `SystemError` class and can have a `path` property in the error object, which contains the file path that caused the error. This is very helpful for `ENOENT` errors. And we can construct a specific user-friendly `ENOENT` error message with it's path.

- Added calls to the new `determineNodeError` function in all switch-cases just to fully ensure they are caught when the exit code is not `null`.

* remove: logging of the node error before the switch-case.

We no longer need to log node errors to the Output channel before the switch-case since we are now logging all errors just before the function rejects the promise and the error is thrown.