fix: prevent the process from hanging if we can't write to `stdin`.
If for some reason `stdin` is null, the process just hangs in the background, without ever closing. This can be seen when the vscode statusbar has the `PHP Sniffer: validating…` text displayed indefinitely.

To fix, we will manually kill the process when the `stdin` is `null`,  which then gracefully resolves the `done` promise callback for the vscode statusbar text.

- Added the `kill` function call, using the `SIGKILL` signal to forcefully terminate the process.

- Added an error message in the `done` promise callback to show the user and log to the Output channel if we did manually kill the process. This is outputted/logged just before it resolves the callback allowing the vscode statusbar text to disappear.