feat: add Windows ENOENT error handling for fixer and sniffer processes
When spawning a command using CMD, and the command is not found, CMD exits with code 1 instead of erroring as ENOENT and then returns the `is not recognized as an internal or external command` message.

*nix systems handle this error correctly by default. So we need to add back the proper ENOENT error in Windows. The cross-spawn package did solve this, but we removed it's usage due to it being unmaintained with issues in newer node versions.

To fix, we can just use the code from cross-spawn under the MIT license, and with major modifications.

- Added new enoent.ts file with functions:

    - `addWindowsEnoentError` function to add the ENOENT error handling. This is the parent function that calls the other functions. This also uses the new `isWin` util function to detect Windows systems.

    - `hookIntoEmit` to hook into the `emit` method of the spawn ChildProcess instance to handle the error.

    - `verifyEnoentError` to verify if we are dealing with an ENOENT error for both spawn and spawnSync.

    - `createEnoentError` to create the ENOENT error and return to the parent function.

- Added `addWindowsEnoentError` function calls to both sniffer and fixer using appropriate params.

- Added a `parsed` object to both sniffer and fixer to pass as argument to `addWindowsEnoentError` function.

This is just to get the ENOENT functions working. The object will eventually come from a separate future function that will escape/quote the command paths/values.