fix: ENOENT handling to use original command info instead of `parsed`.
- Renamed the temporary `parsed` object (introduced in commit 7b62982) to `originalCommand`, and changed it's content to `commandPath` and the original non-parsed `args`.

Also changed all references of the old object and keys to the new ones, in the `windows-enoent-error` file.

This is because the ENOENT error handling only ever uses the `original.args`, `original.command`, and `file` keys, with the latter 2 being the same values. So we can simplify the object. Because the error creation uses the original args, we should also use the non-parsed/non-escaped args.

- Added a proper `existsSync` Node function to the `verifyEnoentError` function to properly check if the path doesn't exist.