fix: warnings should only appear if args were invalid.
When the `validateAdditionalArguments` was refactored in commit d55b7ec, the logger and show user message was moved to the bottom of the function. But an oversight was that this would also warn users even when there were no invalid args.

Fixed by surrounding them in a conditional statement that will only run when there are invalid arguments.