fix: `activateSniffer` shouldn't be checking if sniffer enabled or not.
If sniffer is disabled either via the setting or phpcs executable couldn't be found, the `activateSniffer` previously would return early without finishing setting up some stuff. So when the configuration changed, nothing happened because sniffer didn't get a chance to setup the on change event, and the function only runs on extension startup.

Fixed by removing the conditional statement from the `activateSniffer` function (the check for `snifferEnable` is already in the `validate` function anyway).