fix: remove ending `$` and `m` flag from the regex
The "asserts position at the end of a line" character `$` and the multi-line `m` flag are only needed for formatting issues on the regex101 regex test. The regex for the extension does not require it. It actually breaks the intended purpose of the regex anyway because it asserts to the end of the string which tests negative if there is more text other than the error pattern.

Fixed by removing `$` and `m`.