fix: `normalizePath` to return an empty string if it already is empty.
This fixes issues where if the path is an empty string, then Node's `path.normalize` function will return a string with a dot '.' to indicate cwd.

> If the path is a zero-length string, '.' is returned, representing the current working directory.
~ Node docs - https://nodejs.org/api/path.html#pathnormalizepath

This breaks things, because we allow standard to be an empty string so phpcs can use it's default standard. But phpcs will error saying the '.' is not a standard.

Fixed by allowing the `normalizePath` util function to return an empty string if an empty string is passed to it.