fix: add 'no such file or directory' to PHP not found error regex.
In my GitHub Linux Codespace, I encountered the error: "'php': No such file or directory"
Not sure if this is a real Linux error for php not found, because I changed the name of the php executable to php.bak, just to test. So it could be an erroneous error, but I'll add it just in case.

- Added regex pattern for the  "'php': No such file or directory" error to the `errorPatterns` in `getPhpNotFoundRegex` function.