Fix php not found (#157)
* Revert "Escape filePath and args passed to the phpcs command (#145)"

This reverts commit 58b1febbef5aec408805c4a4f287c31c533e8570.

* fix: spaces in filepaths in both stdpath and standards path

* fix: fixer errors with a 2 double quotes around the paths

Because the fixer doesn't add a `shell: true` to the spawn options, the spawn automatically escapes the command and values, including surrounding them in double quotes. If then we add double quotes here too, the amount of quotes is doubled and errors out:

["ERROR" - 10:33:56 PM] ERROR: the ""d:\Stuart\Websites\Laravel Sites\bedrock-starter\phpcs.xml"" coding standard is not installed. The installed coding standards are MySource, PEAR, PSR1, PSR2, PSR12, Squiz, Zend and JPSR12

* Removes the newline at the end of the fixed code

Closes #138

* fix: 'php' is not recognized" when php isn't available in the PATH.

- Added new `phpExecutablePath` user setting to specify the path to PHP executable, if the user doesn't want to add it to the PATH environment variable.

Cases where not needing php in the PATH are when users are constantly changing PHP versions, and use a version manager like Valet (both macOS and Windows ports).

- Added the functionality to add the user specified PHP executable path to Node's process environment path before running the sniffer and fixer.

- Added an error message to show to the user if the fixer returns with an stderr as "php is not recognized", just to make them aware.

* fix: the check for `php.exe` to account for paths with backslashes as flagged by copilot.

We don't actually need to specify the directory separator. So we can just remove the forward slash.

* fix: TypeScript error (TS18048): `process.env.PATH` is possibly 'undefined' by using the optional chaining operator

* fix: the check for ``php` is not recognized` error being too specific as flagged by copilot.

The original solution only checked for the literal string ``php` is not recognized` but that error message is only on Windows CMD and Powershell.

To fix this and make it OS and shell agnostic, we need to construct a regex and test that instead.

The regex tests for the 5 common formats for the "command not found" error. It also works for with and without single or double quotes around `php`. Also with or without a colon (`:`).

The 5 common formats (without quotes) are:
1. php is not recognized
2. php: command not found
3. php cannot be found
4. command not found: php
5. Unknown command php

While this works for these 5, they're may be some edge cases where certain OS or shells format the error differently and therefore the regex fails. Add those edge cases to the regex as and when they occur.

* feat: add checks for php executable in external extension settings.

- Added new `resolvePhpExecutablePath` settings function to resolve a php executable path.

This adds the ability to check 2 different user settings for the php executable: VSCode's built-in PHP Language Features extension setting `php.validate.executablePath` and Devsense's PHP Tools extension setting `php.executablePath`.

The function acquires the values of these settings plus our new internal setting, and returns the first valid existing path, in the order of precedence from: VSCode built-in extension, Devsense extension, this extension.

- Changed the `phpExecutablePath` value in `loadSettings` function to call the new `resolvePhpExecutablePath` function.

- Added a debug logger to the `addPhpToEnvPath` path resolver util function to log that the php executable has been added to the extension environment.

* build: update the `phpExecutablePath` setting's description.

* refactor: improve regex for detecting 'php not found' error messages

- Removed the multiple redundant `RegExp` object creation calls.

- Simplified the regex and the code building the regex for maintainability.

- Updated the regex101 URL to link to the most recent updated version.

* fix: remove ending `$` and `m` flag from the regex

The "asserts position at the end of a line" character `$` and the multi-line `m` flag are only needed for formatting issues on the regex101 regex test. The regex for the extension does not require it. It actually breaks the intended purpose of the regex anyway because it asserts to the end of the string which tests negative if there is more text other than the error pattern.

Fixed by removing `$` and `m`.

* fix: ts error "Expected 2 arguments, but got 1" for `logger.debug` function

Fix ts error from GitHub Actions: https://github.com/valeryan/vscode-phpsab/actions/runs/19746093177/job/56580557334?pr=157

The error occurred because we were using `logger.debug` function to log a message without the `data` argument.

Fixed by making the `data` argument optional.

* fix: `logger` import path in `path-resolver-utils` to a relative path

Changed the `logger` import path to a relative path since GitHub Actions complains about the `@phpsab` import alias: https://github.com/valeryan/vscode-phpsab/actions/runs/19746225707/job/56580906762?pr=157

* fix: VScode's ts auto module import preferences to `relative`.

Fixed VScode's ts auto module importing by adding the vscode workspace setting `typescript.preferences.importModuleSpecifier` to `relative`.

The setting prevents the `@phpsab' import alias from appearing in the import paths and only relative paths are allowed. By adding it workspace settings, it ensures all maintainers/contributors wil always auto import relatively, which prevents further GitHub Action errors.

* refactor: move the `getPhpNotFoundRegex` to a new `error-helpers` file.

- Created a new `error-helpers` file to house all error functions. We will need this file for v1.0.0 anyway.
- Moved `getPhpNotFoundRegex` function from fixer into the new error-helpers file and added the module import to fixer.

* feat: added the `php is not recognized` error check to sniffer.

- Added the `php is not recognized` error check to sniffer, the same as in fixer - with the `getPhpNotFoundRegex` function and user error message.

* refactor: move `determineNodeError` into `error-helpers`.

- Move `determineNodeError` function from `helpers` into `error-helpers` file for consistency, and change the imports.

* fix: add 'no such file or directory' to PHP not found error regex.

In my GitHub Linux Codespace, I encountered the error: "'php': No such file or directory"
Not sure if this is a real Linux error for php not found, because I changed the name of the php executable to php.bak, just to test. So it could be an erroneous error, but I'll add it just in case.

- Added regex pattern for the  "'php': No such file or directory" error to the `errorPatterns` in `getPhpNotFoundRegex` function.

---------

Co-authored-by: Jonathan Bossenger <jonathanbossenger@gmail.com>
Co-authored-by: Samuel Hilson <samrhilson@gmail.com>