fix: add logging for auto ruleset search and fallback scenarios
- Added logging to the auto ruleset search in `createStandardsPathResolver` function for when:
    - A ruleset is found.
    - A ruleset is not found.

When a ruleset isn't found, the log message could be 1 of 2 messages and is determined whether the `phpsab.standard` setting was set or not. If not set it states it's using the PHPCS default standard, otherwise is says the setting is being used. This log message also shows to the user as a warning pop-up.