fix: exit code 0 is treated as an error in the default case.
Fixes #187

Case 0 in the FIXER's switch-case was removed as redundant in PR #171 (specifically commit d230023). But exit code 0 could be caught in the default case and is treated as an error. Consequently, sending an error notification to the user. This can be seen with the `ignore` option, and the `No violations were found` error message would appear.

This message is synonymous with our `No fixable errors were found` message, and shouldn't be treated as an error.

- Removed the default `message` as `No fixable errors were found.`, this no longer needs to be default and should be set in the switch-case.

- Added case 0, but as a fall-through to case 1, which would be the same code. Falling through cases also prevents duplicate code.

- Added an `else` to the cases 0 & 1 checks to set the message as `no fixable error found`.

- Added a fallback error code message in the default case if the error code couldn't be found in the errors array. This prevents the exit code from returning a description of `undefined`.

The fallback message indicates it's an unknown error and states the error code too for debugging purposes.