fix: add logging for auto ruleset search and fallback scenarios (#186)
* fix: add logging for auto ruleset search and fallback scenarios

- Added logging to the auto ruleset search in `createStandardsPathResolver` function for when:
    - A ruleset is found.
    - A ruleset is not found.

When a ruleset isn't found, the log message could be 1 of 2 messages and is determined whether the `phpsab.standard` setting was set or not. If not set it states it's using the PHPCS default standard, otherwise is says the setting is being used. This log message also shows to the user as a warning pop-up.

* refactor: extract the error code description map into new getter function

- Extracted the code that maps the common error codes to their descriptions from the `determineNodeError` function into a new getter function: `getErrorCodeDescription`.

* fix: add logging for errors during auto ruleset search process.

- Added explicit error logging for when Node's `fs.access` errors out. We catch all errors in an object of objects, with the inner object key being the error code. It has 2 properties, `errorCodeDescription` which uses the new `getErrorCodeDescription` function, and `paths`. For each path tried and failed, it will be pushed in to the `paths` array.

We can then use this object of error objects to construct error messages per error code, and log it the Output channel for debugging.