fix: catch error messages from fixer stdout (#188)
* feat: add new `getEOL` helper function to get the OS end-of-line marker.

- Added `getEOL` helper function to get the OS's end-of-line (EOL) character.

- Added a check to see if the FIXER stdout starts with an EOL character. If it does then log the stdout.

This is important because any PHPCBF error messages sent to stdout always starts with a new line, so by checking for this character, we can test if errors are in the output.

* refactor: extract checks for valid fixed output into new function

- Added new `hasValidFixedOutput` function to check if the FIXER stdout has fixed code (ie. doesn't have error messages or is empty). And added the function calls to cases 1 & 2.

    - Extracted the duplicate checks from the switch-case cases 1 & 2, and added the checks into this function (checks for more than 0 length and different to original file text.
    - Added a check to ensure that the stdout doesn't start with an EOL character (all error messages start with a newline).

* fix: exit code 0 is treated as an error in the default case.

Fixes #187

Case 0 in the FIXER's switch-case was removed as redundant in PR #171 (specifically commit d230023). But exit code 0 could be caught in the default case and is treated as an error. Consequently, sending an error notification to the user. This can be seen with the `ignore` option, and the `No violations were found` error message would appear.

This message is synonymous with our `No fixable errors were found` message, and shouldn't be treated as an error.

- Removed the default `message` as `No fixable errors were found.`, this no longer needs to be default and should be set in the switch-case.

- Added case 0, but as a fall-through to case 1, which would be the same code. Falling through cases also prevents duplicate code.

- Added an `else` to the cases 0 & 1 checks to set the message as `no fixable error found`.

- Added a fallback error code message in the default case if the error code couldn't be found in the errors array. This prevents the exit code from returning a description of `undefined`.

The fallback message indicates it's an unknown error and states the error code too for debugging purposes.